Added Dome Monitor Calibration Function
Added the pixelsWide function in Bassoon/test. This function can be run separately from the GUI in order to figure out the numbers to enter for the psychopy monitor based on measurements. Read function docstring for details. Note, this should only be used when presenting stimuli with Bassoon. Monitor values can change if you present Psychopy stimuli outside of Bassoon.