Added Frame Rate and Removed Window Reset
Experiment object now gets property FR which is the frame rate of its stimulus window - set once.

Took out a section of code that reset the stimulus monitor to a background color in between protocols. Now experiment background value is essentially obsolete, except for brief periods of time at the beginning and end of an experiment. Use the Pause protocols to set a background screen value in between stimuli