Added functionality for adjusting grating means
Protocols: Moving Grating Direction, Oscillating Grating.

The mean intensity of the grating can now be adjusted. Contrast and mean intensity are adjusted separately, but note that the maximum achievable contrast depends on the mean value (e.g. you can achieve higher max contrast if the mean is 0 than if it is 0.75)